Fix circular import in atools for AnalysisMixin
